Villager Updates, Date Decoding, & Name Fixes
This commit adds quite a few new features. Villagers now have most of
their internal structure decoded. When a player's name is changed, they
won't forget you! This commit also incudes a new class, DateUtil.cs,
which allows the decoding of the Animal Crossing date format. You can
see this in action on the Main Form, where the last time the town & each
player played is displayed. Also changed it so when it writes the
player's name, it writes to a few more places, so the game will always
keep track of the player.